# TreasureDeckApi/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "TreasureDeck API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v2/cards": {
            "get": {
                "tags": [
                    "Cards"
                ],
                "summary": "Get list of cards with pagination",
                "operationId": "f7f1f6cc231245966f15aba5ffad75d8",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number for pagination",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Number of items per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of cards",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/CardDTO"
                                            }
                                        },
                                        "meta": {
                                            "type": "object"
                                        },
                                        "links": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Cards"
                ],
                "summary": "Crear una nueva carta",
                "operationId": "76b0337a9b2a3cb1c4147c865a893535",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "collector_number",
                                    "rarity",
                                    "expansion_id"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Black Lotus"
                                    },
                                    "collector_number": {
                                        "type": "string",
                                        "example": "001"
                                    },
                                    "rarity": {
                                        "type": "string",
                                        "example": "Mythic"
                                    },
                                    "expansion_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Carta creada exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CardDTO"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Datos de entrada inválidos"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/api/v2/cards/{id}": {
            "get": {
                "tags": [
                    "Cards"
                ],
                "summary": "Obtener una carta por ID",
                "operationId": "a6f1ba9699e0fd1d7c87746ae19d2755",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la carta",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalle de la carta",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CardDTO"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Carta no encontrada"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Cards"
                ],
                "summary": "Actualizar una carta por ID",
                "operationId": "4de109c9240bbbedb3b42338c935af60",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la carta",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Black Lotus"
                                    },
                                    "collector_number": {
                                        "type": "string",
                                        "example": "001"
                                    },
                                    "rarity": {
                                        "type": "string",
                                        "example": "Mythic"
                                    },
                                    "expansion_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Carta actualizada exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CardDTO"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Carta no encontrada"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Cards"
                ],
                "summary": "Eliminar una carta por ID",
                "operationId": "65409da63cd6eb9e9e0f960d82b0921b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la carta",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Carta eliminada exitosamente"
                    },
                    "404": {
                        "description": "Carta no encontrada"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "CardDTO": {
                "title": "Card",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "name": {
                        "type": "string"
                    },
                    "collector_number": {
                        "type": "string"
                    },
                    "rarity": {
                        "type": "string"
                    },
                    "expansion_id": {
                        "type": "integer"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "BearerAuth": {
                "type": "http",
                "description": "Autenticación usando Bearer token",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}